python-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.18
opendataloader-pdf[hybrid]==1.11.0
openpyxl==3.1.5
python-docx==1.2.0
mammoth==1.11.0
markdownify==0.13.1
python-pptx==1.0.2
requests==2.32.3
